% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{AML_h5ad}
\alias{AML_h5ad}
\title{anndata reference dataset}
\format{
An AnnDataR6 object with 659 features and 250 cells. Included experiments:
\describe{
\item{RNA (main experiment)}{mRNA expression data for 462 genes used as markers for hamatopoietic stem and progenitor (HSPC) cell types at varying stages of differentiation.}
\item{AB}{Surface protein expression data for 197 surface protein markers associated with HSPCs.}
}
}
\usage{
AML_h5ad()
}
\description{
anndata reference dataset
}
\details{
The function described here is used to load the AnnData R6 object, it should be called as a function.
}
