% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{es.mef.small}
\alias{es.mef.small}
\title{Sample data}
\description{
A subset of Saiful et al. 2011 dataset containing first 20 ES and 20 MEF cells.
}
\references{
\url{http://www.ncbi.nlm.nih.gov/pubmed/21543516}
}

