% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{knn}
\alias{knn}
\title{Sample error model}
\description{
SCDE error model generated from the Pollen et al. 2014 dataset.
}
\references{
\url{www.ncbi.nlm.nih.gov/pubmed/25086649}
}

