% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{pagoda.reduce.redundancy}
\alias{pagoda.reduce.redundancy}
\title{Collapse aspects driven by similar patterns (i.e. separate the same sets of cells)}
\usage{
pagoda.reduce.redundancy(tamr, distance.threshold = 0.2,
  cluster.method = "complete", distance = NULL,
  weighted.correlation = TRUE, plot = FALSE, top = Inf, trim = 0,
  abs = FALSE, ...)
}
\arguments{
\item{tamr}{output of pagoda.reduce.loading.redundancy()}

\item{distance.threshold}{similarity threshold for grouping interdependent aspects}

\item{cluster.method}{one of the standard clustering methods to be used (fastcluster::hclust is used if available or stats::hclust)}

\item{distance}{}

\item{weighted.correlation}{Boolean of whether to use a weighted correlation in determining the similarity of patterns}

\item{plot}{Boolean of whether to show plot}

\item{top}{Restrict output to the top n aspects of heterogeneity}

\item{trim}{Winsorization trim to use prior to determining the top aspects}

\item{abs}{Boolean of whether to use absolute correlation}

\item{...}{additional arguments are passed to the pagoda.view.aspects() method during plotting}

\item{corr.power}{power to which the product of loading and score correlation is raised}
}
\value{
a list structure analogous to that returned by pagoda.top.aspects(), but with addition of a $cnam element containing a list of aspects summarized by each row of the new (reduced) $xv and $xvw
}
\description{
Examines PC loading vectors underlying the identified aspects and clusters aspects based on score correlation. Clusters of aspects driven by the same patterns are determined based on the distance.threshold.
}

