% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{pagoda.top.aspects}
\alias{pagoda.top.aspects}
\title{Score statistical significance of gene set and cluster overdispersion}
\usage{
pagoda.top.aspects(pwpca, clpca = NULL, n.cells = NULL,
  z.score = qnorm(0.05/2, lower.tail = F), return.table = F,
  return.genes = F, plot = F, adjust.scores = T, score.alpha = 0.05,
  use.oe.scale = F, effective.cells.start = NULL)
}
\arguments{
\item{pwpca}{output of pagoda.pathway.wPCA()}

\item{clpca}{output of pagoda.gene.clusters() (optional)}

\item{n.cells}{effective number of cells (if not provided, will be determined using pagoda.effective.cells())}

\item{z.score}{Z score to be used as a cutoff for statistically significant pattenrs (defaults to 0.05 P-value}

\item{return.table}{whether a text table showing}

\item{return.genes}{whether a set of genes driving significant aspects should be returned}

\item{plot}{whether to plot the cv/n vs. dataset size scatter showing significance models}

\item{adjust.scores}{whether the normalization of the aspect patterns should be based on the adjusted Z scores - qnorm(0.05/2,lower.tail=F)}

\item{score.alpha}{significance lelvel of the confidence interval for determining upper/lower bounds}

\item{use.oe.scale}{whether the variance of the returned aspect patterns should be normalized using oberved/expected value instead of the default chi-squared derived variance corresponding to overdispersion Z score}

\item{effective.cells.start}{starting value for the pagoda.effective.cells() call}
}
\value{
if return.table=F and return.genes=F (default) returns a list structure containing the following items:
\itemize{
\item{xv} {a matrix of normalized aspect patterns (rows- significant aspects, columns- cells}
\item{xvw} { corresponding weight matrix }
\item{gw} { set of genes driving the significant aspects }
\item{df} { text table with the significance testing results }
}
}
\description{
Evaluates statistical significance of the gene set and cluster lambda1 values, returning
either a text table of Z scores, etc, a structure containing normalized values of significant
aspects, or a set of genes underlying the significant aspects.
}

