% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{pagoda.view.aspects}
\alias{pagoda.view.aspects}
\title{View PAGODA output}
\usage{
pagoda.view.aspects(tamr, row.clustering = hclust(dist(tamr$xv)), top = Inf,
  ...)
}
\arguments{
\item{tamr}{Combined pathways that show similar expression patterns. Output of \code{\link{pagoda.reduce.redundancy}}}

\item{row.clustering}{Dendrogram of combined pathways clustering}

\item{top}{Restrict output to the top n aspects of heterogeneity}

\item{...}{additional arguments are passed to the \code{\link{view.aspects}} method during plotting}
}
\value{
PAGODA heatmap
}
\description{
Create static image of PAGODA output visualizing cell hierarchy and top aspects of transcriptional heterogeneity
}

