% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{papply}
\alias{papply}
\title{wrapper around different mclapply mechanisms}
\usage{
papply(..., n.cores = n)
}
\arguments{
\item{...}{parameters to pass to lapply, mclapply, bplapply, etc.}

\item{n.cores}{number of cores. If 1 core is requested, will default to lapply}
}
\description{
Abstracts out mclapply implementation, and defaults to lapply when only one core is requested (helps with debugging)
}

