% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDE-package.R
\docType{package}
\name{scDE-package}
\alias{scDE}
\alias{scDE-package}
\title{scDE: Unified Single Cell Differential Expression}
\description{
Tools for performing differential expression analysis in single cell objects and returning common results. scDE currently accepts Seurat and Anndata objects.
}
\author{
\strong{Maintainer}: William Showers \email{bill.showers@refinedscience.com}

Authors:
\itemize{
  \item Jairav Desai \email{jai.desai@refinedscience.com}
  \item Austin Gillen \email{austin.gillen@refinedscience.com}
}

}
\keyword{internal}
