% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{scde.browse.diffexp}
\alias{scde.browse.diffexp}
\title{View differential expression results in a browser}
\usage{
scde.browse.diffexp(results, models, counts, prior, groups = NULL,
  batch = NULL, geneLookupURL = NULL, server = NULL, name = "scde",
  port = NULL)
}
\arguments{
\item{results}{result object returned by \code{scde.expression.difference()}. Note to browse group posterior levels, use \code{return.posteriors = TRUE} in the \code{scde.expression.difference()} call.}

\item{models}{model matrix}

\item{counts}{count matrix}

\item{prior}{prior}

\item{groups}{group information}

\item{batch}{batch information}

\item{geneLookupURL}{The URL that will be used to construct links to view more information on gene names. By default (if can't guess the organism) the links will forward to ENSEMBL site search, using \code{geneLookupURL = "http://useast.ensembl.org/Multi/Search/Results?q = {0}"}. The "{0}" in the end will be substituted with the gene name. For instance, to link to GeneCards, use \code{"http://www.genecards.org/cgi-bin/carddisp.pl?gene = {0}"}.}

\item{server}{optional previously returned instance of the server, if want to reuse it.}

\item{name}{app name (needs to be altered only if adding more than one app to the server using \code{server} parameter)}

\item{port}{Interactive browser port}
}
\value{
server instance, on which $stop() function can be called to kill the process.
}
\description{
Launches a browser app that shows the differential expression results, allowing to sort, filter, etc.
The arguments generally correspond to the \code{scde.expression.difference()} call, except that the results of that call are also passed here. Requires \code{Rook} and \code{rjson} packages to be installed.
}
\examples{
\donttest{
data(es.mef.small)
cd <- es.mef.small
cd <- cd[rowSums(cd) > 0, ]
cd <- cd[, colSums(cd) > 1e4]
sg <- factor(gsub("(MEF|ESC).*", "\\\\1", colnames(cd)), levels = c("ESC", "MEF"))
names(sg) <- colnames(cd)
o.ifm <- scde.error.models(counts = cd, groups = sg, n.cores = 10, threshold.segmentation = TRUE)
o.prior <- scde.expression.prior(models = o.ifm, counts = cd, length.out = 400, show.plot = FALSE)
# make sure groups corresponds to the models (o.ifm)
groups <- factor(gsub("(MEF|ESC).*", "\\\\1", rownames(o.ifm)), levels = c("ESC", "MEF"))
names(groups) <- row.names(o.ifm)
ediff <- scde.expression.difference(o.ifm, cd, o.prior, groups = groups, n.randomizations = 100, n.cores = 10, verbose = 1)
scde.browse.diffexp(ediff, o.ifm, cd, o.prior, groups = groups, geneLookupURL="http://www.informatics.jax.org/searchtool/Search.do?query={0}")  # creates browser
}
}

