% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{scde.error.models}
\alias{scde.error.models}
\title{Fit single-cell error/regression models}
\usage{
scde.error.models(counts, groups = NULL, min.nonfailed = 3,
  threshold.segmentation = T, min.count.threshold = 4,
  zero.count.threshold = min.count.threshold, zero.lambda = 0.1,
  save.crossfit.plots = F, save.model.plots = T, n.cores = 12,
  min.size.entries = 2000, max.pairs = 5000, min.pairs.per.cell = 10,
  verbose = 0, linear.fit = F, local.theta.fit = T,
  theta.fit.range = c(0.01, 100))
}
\arguments{
\item{counts}{read count matrix. The rows correspond to genes (should be named), columns correspond to individual cells. The matrix should contain integer counts}

\item{groups}{an optional factor describing grouping of different cells. If provided, the cross-fits and the expected expression magnitudes will be determined separately within each group. The factor should have the same length as ncol(counts).}

\item{min.nonfailed}{minimal number of non-failed observations required for a gene to be used in the final model fitting}

\item{threshold.segmentation}{use a fast threshold-based segmentation during cross-fit (default: T)}

\item{min.count.threshold}{the number of reads to use to guess which genes may have "failed" to be detected in a given measurement during cross-cell comparison (default 4 reads)}

\item{zero.count.threshold}{threshold to guess the initial value (failed/non-failed) during error model fitting procedure (defaults to the min.count.threshold value)}

\item{zero.lambda}{the rate of the Poisson (failure) component (default : 0.1)}

\item{save.crossfit.plots}{whether png files showing cross-fit segmentations should be written out (default: F)}

\item{save.model.plots}{whether pdf files showing model fits should be written out (default = T)}

\item{n.cores}{number of cores to use}

\item{min.size.entries}{minimum number of genes to use when determining expected expression magnitude during model fitting}

\item{max.pairs}{maximum number of cross-fit comparisons that should be performed per group (default 5000)}

\item{min.pairs.per.cell}{minimum number of pairs that each cell should be cross-compared with}

\item{verbose}{1 for increased output}
}
\value{
a model matrix, with rows corresponding to different cells, and columns representing different parameters of the determined models

a data frame containing model parameters
}
\description{
Fit error models given a set of single-cell data (counts) and an optional grouping factor (groups). The cells (within each group) are first cross-compared to determine a subset of genes showing consistent expression. The set of genes is then used to fit a mixture model (Poisson-NB mixture, with expression-dependent concomitant).
}
\examples{
\donttest{
data(es.mef.small)
cd <- es.mef.small cd <- cd[rowSums(cd) > 0, ] cd <- cd[, colSums(cd) > 1e4]
sg <- factor(gsub("(MEF|ESC).*", "\\\\1", colnames(cd)), levels = c("ESC", "MEF")) names(sg) <- colnames(cd)
o.ifm <- scde.error.models(counts = cd, groups = sg, n.cores = 10, threshold.segmentation = T)
}
}

