% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{scde.expression.magnitude}
\alias{scde.expression.magnitude}
\title{Return scaled expression magnitude estimates}
\usage{
scde.expression.magnitude(models, counts)
}
\arguments{
\item{models}{models determined by \code{\link{scde.error.models}}}

\item{counts}{count matrix}
}
\value{
a matrix of expression magnitudes on a log scale (rows - genes, columns - cells)
}
\description{
Return point estimates of expression magnitudes of each gene across a set of cells, based on the regression slopes determined during the model fitting procedure.
}
\examples{
\donttest{
data(es.mef.small)
cd <- es.mef.small
cd <- cd[rowSums(cd) > 0, ]
cd <- cd[, colSums(cd) > 1e4]
sg <- factor(gsub("(MEF|ESC).*", "\\\\1", colnames(cd)), levels = c("ESC", "MEF"))
names(sg) <- colnames(cd)
o.ifm <- scde.error.models(counts = cd, groups = sg, n.cores = 10, threshold.segmentation = T)
# get expression magniutude estimates
lfpm <- scde.expression.magnitude(o.ifm, cd)
}
}

