% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{winsorize.matrix}
\alias{winsorize.matrix}
\title{Winsorize matrix}
\usage{
winsorize.matrix(mat, trim)
}
\arguments{
\item{mat}{matrix}

\item{trim}{fraction of outliers (on each side) that should be Winsorized, or (if the value is >=1) the number of outliers to be trimmed on each side}
}
\value{
Winsorized matrix
}
\description{
Sets the ncol(mat)*trim top outliers in each row to the next lowest value; same for the lowest outliers
}

