% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{pagoda.cluster.cells}
\alias{pagoda.cluster.cells}
\title{Determine optimal cell clustering based on the genes driving the significant aspects}
\usage{
pagoda.cluster.cells(tam, varinfo, method = "ward", include.aspects = F,
  verbose = 0, return.details = F)
}
\arguments{
\item{tam}{result of pagoda.top.aspects() call}

\item{varinfo}{result of pagoda.varnorm() call}

\item{method}{clustering method ('ward' by default)}

\item{include.aspects}{whether the aspect patterns themselves should be included alongside with the individual genes in calculating cell distance}
}
\value{
hclust result
}
\description{
Determines cell clustering (hclust result) based on a weighted correlation of genes
underlying the top aspects of transcriptional heterogeneity. Branch orientation is optimized
if 'cba' package is installed.
}

