% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dge.R
\name{run_dge}
\alias{run_dge}
\alias{run_dge.Seurat}
\alias{run_dge.AnnDataR6}
\title{Differential Gene Expression}
\usage{
run_dge(object, group_by, lfc_format = "default", ...)

\method{run_dge}{Seurat}(object, group_by, lfc_format = "default")

\method{run_dge}{AnnDataR6}(object, group_by, lfc_format = "default")
}
\arguments{
\item{object}{a single-cell object. Currently, Seurat and
SingleCellExperiment objects are supported.}

\item{group_by}{metadata variable to use for forming differential gene
expression groups. A group will be created for all values in this variable
present in the object passed to this function.}

\item{lfc_format}{sets the format used for reporting log-fold changes. If
"default", the default formats for each test used will be used. If "log2",
log-fold changes will be reported in a log two fold change format, and if
"ln", log-fold changes will be reported as a natural log fold change.}
}
\description{
Performs differential expression testing on single cell data depending on the
class of the object passed.
}
\details{
Tests used
- Seurat objects: Presto
- Anndata objects: rank_genes_groups, using the "wilcoxon" method
}
\section{Methods (by class)}{
\itemize{
\item \code{run_dge(Seurat)}: Seurat objects: uses Presto

\item \code{run_dge(AnnDataR6)}: Anndata objects: uses Scanpy's rank_genes_groups
function, with the "wilcoxon" method

}}
