% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dge.R
\name{run_dge}
\alias{run_dge}
\alias{run_dge.Seurat}
\alias{run_dge.AnnDataR6}
\title{Differential Gene Expression}
\usage{
run_dge(
  object,
  group_by,
  seurat_assay = NULL,
  slot = NULL,
  lfc_format = "default",
  positive_only = FALSE,
  remove_raw_pval = FALSE,
  ...
)

\method{run_dge}{Seurat}(
  object,
  group_by,
  seurat_assay = NULL,
  slot = NULL,
  lfc_format = "default",
  positive_only = FALSE,
  remove_raw_pval = FALSE
)

\method{run_dge}{AnnDataR6}(
  object,
  group_by,
  seurat_assay = NULL,
  slot = NULL,
  lfc_format = "default",
  positive_only = FALSE,
  remove_raw_pval = FALSE
)
}
\arguments{
\item{object}{a single-cell object. Currently, Seurat and
SingleCellExperiment objects are supported.}

\item{group_by}{metadata variable to use for forming differential gene
expression groups. A group will be created for all values in this variable
present in the object passed to this function.}

\item{seurat_assay}{for Seurat objects, the assay to use for evaluating
differences in expression. This is used only for tests on Seurat objects.
For SingleCellExperiment objects, the assay defaults to the main experiment,
and for anndata objects the "X" matrix is used.}

\item{slot}{The feature matrix slot to pull data from. If NULL, the "data"
slot is used for Seurat objects, and the "logcounts" slot is used for
SingleCellExperiment objects. This parameter is currently ignored for
anndata objects: only the "X" matrix may be used.}

\item{lfc_format}{sets the format used for reporting log-fold changes. If
"default", the default formats for each test used will be used. If "log2",
log-fold changes will be reported in a log two fold change format, and if
"ln", log-fold changes will be reported as a natural log fold change.}

\item{positive_only}{If TRUE, only display genes that are up-regulated in a
group compared to the reference (LFC > 0). Defaults to FALSE.}

\item{remove_raw_pval}{If TRUE, show only the adjusted p-value.
Defaults to FALSE.}
}
\description{
Performs differential expression testing on single cell data depending on the
class of the object passed. See the "Methods (by class)" section for details
on the test used.
}
\section{Methods (by class)}{
\itemize{
\item \code{run_dge(Seurat)}: Seurat objects: uses Presto

\item \code{run_dge(AnnDataR6)}: Anndata objects: uses Scanpy's rank_genes_groups
function, with the "wilcoxon" method

}}
