% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{scde}
\alias{scde}
\alias{scde-package}
\title{Single-cell Differential Expression (with Pathway And Gene set Overdispersion Analysis)}
\description{
The scde package implements a set of statistical methods for analyzing single-cell RNA-seq data.
scde fits individual error models for single-cell RNA-seq measurements. These models can then be used for
assessment of differential expression between groups of cells, as well as other types of analysis.
The scde package also contains the pagoda framework which applies pathway and gene set overdispersion analysis
to identify and characterize putative cell subpopulations based on transcriptional signatures.
See vignette("diffexp") for a brief tutorial on differential expression analysis.
See vignette("pagoda") for a brief tutorial on pathway and gene set overdispersion analysis to identify and characterize cell subpopulations.
More extensive tutorials are available at \url{http://pklab.med.harvard.edu/scde/index.html}.
 (test)
}
\author{
Peter Kharchenko \email{Peter_Kharchenko@hms.harvard.edu}

Jean Fan \email{jeanfan@fas.harvard.edu}
}

